refactor(WinningMoney.java) : 패키지 구조 변경 당첨 / 로또 패키지 구분 X
