+++ docs/README.md
@@ -83,8 +83,6 @@
 - [x] 로또 번호들 리스트를 생성한다.
   - [x] 구매 횟수가 0보다 커야한다.
   - [x] 구매 횟수 만큼의 로또 번호 리스트를 생성한다.
-
-### 당첨
 #### 당첨 번호
 - [x] 당첨 번호를 생성.
   - [x] 숫자만 입력 가능하다.
@@ -101,10 +99,10 @@
 - [x] 당첨 금액 생성.
 - [x] 일치 갯수별 담청 금액을 계산
   - [x] 일치 갯수를 만족하는 로또 갯수 * 일치 갯수별 당첨 금액
+#### 통계
 - [ ] 수익률을 계산.
   - [ ] 구매 구입 금액에서 당첨 금액을 차감한 후 구매 금액으로 나눈다.
   - [ ] 소수점 두자리
-
 - [ ] 컴포지트 패턴 winning number
 - [ ] enum null 제거 / stream
 - [ ]0 제거 lottoGenerator
