--- src/main/java/common/vo/Count.java
@@ -1,29 +1,0 @@
-package common.vo;
-
-import java.util.Objects;
-
-public class Count {
-
-    private int number;
-
-    public Count(int number) {
-        this.number = number;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Count count = (Count) o;
-        return number == count.number;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number);
-    }
-
-    public int getNumber() {
-        return this.number;
-    }
-}
