+++ src/main/java/lotto/domain/Count.java
@@ -1,0 +1,29 @@
+package lotto.domain;
+
+import java.util.Objects;
+
+public class Count {
+
+    private int number;
+
+    public Count(int number) {
+        this.number = number;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Count count = (Count) o;
+        return number == count.number;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(number);
+    }
+
+    public int getNumber() {
+        return this.number;
+    }
+}
