+++ src/main/java/lotto/ui/inputView/WinningNumberInputView.java
@@ -1,0 +1,12 @@
+package lotto.ui.inputView;
+
+import java.util.Scanner;
+
+public class WinningNumberInputView {
+
+    public static String readWinningNumbers() {
+        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
+        Scanner s = new Scanner(System.in);
+        return s.nextLine();
+    }
+}
