+++ src/test/java/lotto/fixture/LottoMoneyFixture.java
@@ -1,0 +1,9 @@
+package lotto.fixture;
+
+import lotto.domain.LottoMoney;
+
+public class LottoMoneyFixture {
+    public static LottoMoney 로또구매금액_10_000() {
+        return new LottoMoney(10_000);
+    }
+}
