--- src/test/java/lotto/lotto/domain/fixture/LottoMoneyFixture.java
@@ -1,9 +1,0 @@
-package lotto.lotto.domain.fixture;
-
-import lotto.lotto.domain.LottoMoney;
-
-public class LottoMoneyFixture {
-    public static LottoMoney 로또구매금액_10_000() {
-        return new LottoMoney(10_000);
-    }
-}
